ecommerce/services/reviews/requirements.txt
flask
psycopg2-binary
PyJWT==2.6.0